Backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
gunicorn==20.1.0
